requirements.txt
discord.py
python-dotenv~=0.15.0
unidecode~=1.1.2
googletrans~=4.0.0rc1
pyowm~=3.1.1
yt_dlp
